Rename timer to alarm
To match the Tock drivers rename the libtock-rs timer to alarm.
