+++ examples/blink.rs
@@ -1,7 +1,7 @@
 #![no_std]
 
+use libtock::alarm::Duration;
 use libtock::result::TockResult;
-use libtock::timer::Duration;
 
 libtock_core::stack_size! {0x400}
 
@@ -10,7 +10,7 @@ async fn main() -> TockResult<()> {
     let mut drivers = libtock::retrieve_drivers()?;
 
     let leds_driver = drivers.leds.init_driver()?;
-    let mut timer_driver = drivers.timer.create_timer_driver();
+    let mut timer_driver = drivers.alarm.create_timer_driver();
     let timer_driver = timer_driver.activate()?;
 
     // Blink the LEDs in a binary count pattern and scale
