+++ examples/gpio.rs
@@ -1,7 +1,7 @@
 #![no_std]
 
+use libtock::alarm::Duration;
 use libtock::result::TockResult;
-use libtock::timer::Duration;
 
 libtock_core::stack_size! {0x800}
 
@@ -11,7 +11,7 @@ async fn main() -> TockResult<()> {
     let mut drivers = libtock::retrieve_drivers()?;
 
     let mut gpio_driver = drivers.gpio.init_driver()?;
-    let mut timer_driver = drivers.timer.create_timer_driver();
+    let mut timer_driver = drivers.alarm.create_timer_driver();
     let timer_driver = timer_driver.activate()?;
 
     let mut gpio = gpio_driver.gpios().next().unwrap();
