+++ examples/sensors.rs
@@ -1,9 +1,9 @@
 #![no_std]
 
+use libtock::alarm::Duration;
 use libtock::println;
 use libtock::result::TockResult;
 use libtock::sensors::Sensor;
-use libtock::timer::Duration;
 
 libtock_core::stack_size! {0x800}
 
@@ -11,7 +11,7 @@ libtock_core::stack_size! {0x800}
 async fn main() -> TockResult<()> {
     let mut drivers = libtock::retrieve_drivers()?;
 
-    let mut timer_driver = drivers.timer.create_timer_driver();
+    let mut timer_driver = drivers.alarm.create_timer_driver();
     let timer_driver = timer_driver.activate()?;
     drivers.console.create_console();
