+++ examples/timer_parallel.rs
@@ -1,10 +1,10 @@
 #![no_std]
 
 use futures::future;
+use libtock::alarm::Duration;
+use libtock::alarm::ParallelSleepDriver;
 use libtock::leds::Led;
 use libtock::result::TockResult;
-use libtock::timer::Duration;
-use libtock::timer::ParallelSleepDriver;
 
 libtock_core::stack_size! {0x800}
 
@@ -25,7 +25,7 @@ async fn main() -> TockResult<()> {
     let mut drivers = libtock::retrieve_drivers()?;
 
     let leds_driver = drivers.leds.init_driver()?;
-    let mut timer_driver = drivers.timer.create_timer_driver();
+    let mut timer_driver = drivers.alarm.create_timer_driver();
     let timer_driver = timer_driver.activate()?;
 
     let mut leds = leds_driver.leds();
